- Fixed potential bug with texture internal type size

+++ engine/source/engine/render/texture_manager/texture_mng.h
@@ -172,8 +172,8 @@ private:
 private:
     ds::StrID m_name = "";
     
-    uint16_t m_type = 0;
-    uint16_t m_levelsCount = 0;
+    uint32_t m_type = 0;
+    uint32_t m_levelsCount = 0;
     
     uint32_t m_width = 0;
     uint32_t m_height = 0;
